+++ projects/biomegen/script.js
@@ -5,13 +5,12 @@ const startTime = new Date();
 // Test Result Class
 
 class TestResult {
-    constructor(version, width, height, processes, mean, std_dev, pix_per_sec, pct5, pct50, pct95){
+    constructor(version, width, height, processes, mean, pix_per_sec, pct5, pct50, pct95){
         this.version = version;
         this.width = width;
         this.height = height;
         this.processes = processes;
         this.mean = mean;
-        this.std_dev = std_dev;
         this.pix_per_sec = pix_per_sec;
         this.pct5 = pct5;
         this.pct50 = pct50;
@@ -42,8 +41,8 @@ for (let i in csvTextLines) {
         new TestResult(
             textLine[0], Number(textLine[1]), Number(textLine[2]), // version, width, height
             Number(textLine[3]), // processes (reps ignored)
-            Number(textLine[5]), Number(textLine[6]), // mean, standard_deviation
-            Number(textLine[7]), // pixels per second
+            Number(textLine[5]), Number(textLine[7]),
+            // mean (standard deviation ignored), pixels per second
             Number(textLine[8]), Number(textLine[10]), Number(textLine[12])
             // 5th, 50th, and 95th percentiles (25th and 75th ignored)
         )
@@ -51,6 +50,19 @@ for (let i in csvTextLines) {
 }
 
 const newestVersion = testResults[testResults.length - 1].version;
+
+let bestProcesses = {1920: 8, 2560: 8, 3840: 8, 7680: 8, 10000: 8};
+let results8 = {1920: 0, 2560: 0, 3840: 0, 7680: 0, 10000: 0};
+for (let i in testResults) {
+    let result = testResults[i];
+    if (result.version == newestVersion) {
+        if (result.processes == 8) {
+            results8[result.width] = result.mean;
+        } else if (result.processes == 16 && result.mean < results8[result.width]) {
+            bestProcesses[result.width] == 16;
+        }
+    }
+}
 
 // Getting Graph Values
 
@@ -62,15 +74,13 @@ let xValues = {
 };
 let yValues = {
     "Version vs Time": [[], [], [], []], // mean, 5th pct, 50th pct, 95th pct
-    "Version vs Pix Per Sec": [[], [], [], [], []], // 1080p, 1440p, 4K
-    "Version vs Std Dev": [[], [], [], [], []],
+    "Version vs Pix Per Sec": [[], [], [], [], []], // 1080p, 1440p, 4K, 8K, 10K
     "Resolution vs Time": [[], [], [], []],
     "Pixels vs Time": [[], [], [], []],
     "Resolution vs Pix Per Sec": [],
     "Pixels vs Pix Per Sec": [],
     "Processes vs Time": [[], [], [], []],
-    "Processes vs Pix Per Sec": [[], [], [], [], []],
-    "Processes vs Std Dev": [[], [], [], [], []]
+    "Processes vs Pix Per Sec": [[], [], [], [], []]
 };
 
 for (let i in testResults) {
@@ -80,8 +90,8 @@ for (let i in testResults) {
     // Not in order
 
     if (
-        (result.version != "3.0.1" && result.processes === 8) ||
-        (result.version === "3.0.1" && result.processes === 16)
+        result.version != newestVersion ||
+        (result.version == newestVersion && (bestProcesses[result.width] == result.processes))
     ) {
 
         let index = 0;
@@ -126,10 +136,6 @@ for (let i in testResults) {
         }
 
         yValues["Version vs Pix Per Sec"][index].push(result.pix_per_sec);
-
-        yValues["Version vs Std Dev"][index].push(
-            result.std_dev / result.mean * 100
-        );
 
         if (result.version === newestVersion) {
 
@@ -198,10 +204,6 @@ for (let i in testResults) {
         }
 
         yValues["Processes vs Pix Per Sec"][index].push(result.pix_per_sec);
-
-        yValues["Processes vs Std Dev"][index].push(
-            result.std_dev / result.mean * 100
-        );
 
     }
 
@@ -324,62 +326,6 @@ new Chart(graph, {
         }
     }
 });
-
-// Processes vs Standard Deviation
-
-graph = document.getElementById("graph-version-std_dev");
-
-new Chart(graph, {
-    type: "line",
-    data: {
-        labels: xValues["Version"],
-        datasets: [
-            {
-                label: "1080p",
-                data: yValues["Version vs Std Dev"][0]
-            },
-            {
-                label: "1440p",
-                data: yValues["Version vs Std Dev"][1]
-            },
-            {
-                label: "4K",
-                data: yValues["Version vs Std Dev"][2]
-            },
-            {
-                label: "8K",
-                data: yValues["Version vs Std Dev"][3]
-            },
-            {
-                label: "10K x 10K",
-                data: yValues["Version vs Std Dev"][4]
-            }
-        ]
-    },
-    options: {
-        plugins: {
-            legend: {
-                display: true,
-                position: "bottom",
-            },
-            title: {
-                display: true,
-                text: "Version vs Standard Deviation",
-                font: {size: 26}
-            }
-        },
-        scales: {
-            x: {
-                grid: {color: "#606060"}
-            },
-            y: {
-                grid: {color: "#606060"},
-                min: 0
-            }
-        }
-    }
-});
-
 
 // Resolution vs Time
 
@@ -656,61 +602,6 @@ new Chart(graph, {
             title: {
                 display: true,
                 text: "Processes vs Pixels per Second",
-                font: {size: 26}
-            }
-        },
-        scales: {
-            x: {
-                grid: {color: "#606060"}
-            },
-            y: {
-                grid: {color: "#606060"},
-                min: 0
-            }
-        }
-    }
-});
-
-// Processes vs Standard Deviation
-
-graph = document.getElementById("graph-processes-std_dev");
-
-new Chart(graph, {
-    type: "line",
-    data: {
-        labels: xValues["Processes"],
-        datasets: [
-            {
-                label: "1080p",
-                data: yValues["Processes vs Std Dev"][0]
-            },
-            {
-                label: "1440p",
-                data: yValues["Processes vs Std Dev"][1]
-            },
-            {
-                label: "4K",
-                data: yValues["Processes vs Std Dev"][2]
-            },
-            {
-                label: "8K",
-                data: yValues["Processes vs Std Dev"][3]
-            },
-            {
-                label: "10K x 10K",
-                data: yValues["Processes vs Std Dev"][4]
-            }
-        ]
-    },
-    options: {
-        plugins: {
-            legend: {
-                display: true,
-                position: "bottom",
-            },
-            title: {
-                display: true,
-                text: "Processes vs Standard Deviation (Percent of Mean Time)",
                 font: {size: 26}
             }
         },